fix(Page): rename Head to Helmet (#2079)
