+++ .storybook/stories/tutorials/Layout/story/index.jsx
@@ -93,9 +93,9 @@ which is a part of Picasso librabry. It should fill all available height of the 
 tutorialChapter
   .addTextSection(
     `
-Main content can have various layouts, but for this tutorial, we choose to have a title 
-and two sections. We use [\`PageHead\`](..?path=/story/picasso-lab-pagehead--pagehead) for 
-the title and [\`Section\`](..?path=/story/picasso-lab-section--section) for content 
+Main content can have various layouts, but for this tutorial, we choose to have a title
+and two sections. We use [\`PageHead\`](..?path=/story/picasso-lab-pagehead--pagehead) for
+the title and [\`Section\`](..?path=/story/picasso-lab-section--section) for content
 sections. Also, it's very important to use use
 \`Page.Article\` component, because we use a flexbox wrapper
 around the page and we need to make \`MainContent\` grow and fill all available space.
