--- packages/picasso/src/PageHead/index.ts
@@ -1,7 +1,0 @@
-import { OmitInternalProps } from '@toptal/picasso-shared'
-
-import { Props } from './PageHead'
-
-export { default } from './PageHead'
-export type PageHeadProps = OmitInternalProps<Props>
-export * from './PageHead'
