+++ packages/picasso/src/PageHelmet/index.ts
@@ -1,0 +1,7 @@
+import { OmitInternalProps } from '@toptal/picasso-shared'
+
+import { Props } from './PageHelmet'
+
+export { default } from './PageHelmet'
+export type PageHelmetProps = OmitInternalProps<Props>
+export * from './PageHelmet'
